add load hook
and I think I improved the load time (can't say for sure). At least it loads faster than the way I was going to do it (pathlib sucks when it comes to speed).

Now you can show the loading progress in a gui. I mainly just added this feature because I want to be able to show a progress bar when loading a game in wme.