add dump function to Filesystem
It just dumps all the files into the specified directory (or in the assets directory that the game was loaded from), so now all changes to files can get saved. I just need to add a system to only dump files that were edited, and a system to save the original file location of a file. That would make it much easier to edit and save only one file.