Fix numbered properties in `ready_properties()`
All numbered properties would be removed if they were equal to the default property, regardless of whether it was specified to be included or not.

I also decided to add some default properties to the object if there were no level properties on object creation (when they're added to a level).